re-organized lexer and parser
